Added code headers to files where missing, added documentation, and renamed some functions
