+++ libs/rpc/src/client/local.test.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 import { describe, expect, test } from "vitest"
 import { handleLocalModuleMethod } from "./local"
 import type { PrimOptions } from "../interfaces"
