+++ libs/rpc/src/client/local.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 import { givenFormLike, handlePossibleForm } from "../extract/blobs"
 import type { PossibleModule } from "../interfaces"
 import type { UserProvidedClientOptions } from "../options/client/provided"
