+++ libs/rpc/src/client/proxy.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 import { DeepProxy } from "proxy-deep"
 import { nanoid } from "nanoid"
 import type { RpcCall, RpcChain } from "../types/rpc-structure"
