+++ libs/rpc/src/client/wrapper.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 /** The default behavior upon receiving an error in `handlePotentialPromise()` is to rethrow that error */
 function defaultError(error: unknown) {
 	throw error
