+++ libs/rpc/src/constants.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 import { nanoid } from "nanoid"
 
 /** Placeholder names for extracted values from RPC */
@@ -7,10 +11,12 @@ export enum RpcPlaceholder {
 	PromisePrefix = "prom",
 }
 
-export function placeholderName(type: RpcPlaceholder, id?: string) {
+/** Generate an identifier with the given prefix and, optionally, identifier (default is randomly generated) */
+export function generateId(type: RpcPlaceholder, id?: string) {
 	return `_${type}_${id ?? nanoid()}`
 }
 
-export function placeholderOnly(given: RpcPlaceholder) {
-	return placeholderName(given, "")
+/** Used to place `given` prefix into prefix format: `_${given}_` _without_ an identifier */
+export function resolvePlaceholder(given: RpcPlaceholder) {
+	return generateId(given, "")
 }
