+++ libs/rpc/src/error.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 enum RpcErrorCode {
 	InvalidRpcResult = 0,
 }
