+++ libs/rpc/src/extract/base.test.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 import { expect, test } from "vitest"
 import { extract, iterate, merge } from "./base"
