+++ libs/rpc/src/types/rpc-module.test-d.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 import { expectTypeOf, test, describe } from "vitest"
 import type { RpcModule } from "./rpc-module"
