+++ libs/rpc/src/types/rpc-structure.ts
@@ -1,3 +1,7 @@
+// Part of the Prim+RPC project ( https://prim.doseofted.me/ )
+// Copyright 2023 Ted Klingenberg
+// SPDX-License-Identifier: Apache-2.0
+
 export interface RpcBase {
 	prim?: 0 // as new major versions are announced (only that change message structure), bump version
 	id?: string | number
